# [RyuJIT] Avoid emitting cdq/coq for a positive constant in rax for idiv